Simple_flask_app_using_docker/requirements.txt
numpy
scipy
qtpy
psutil
Flask
flask-wtf
flask-restful
gunicorn
sqlalchemy
requests
pyyaml
ruamel.yaml